Organizing tooltip interface and data in separate files and binding tooltipview
